Updated downloader class
(1) Prompt before deleting the download directory  
(2) Alow for downloading specific class videos or downloading videos from a range of classes.

+++ KineticsDatasetManager.py
@@ -31,8 +31,6 @@ Purpose:	Manage kinetics dataset according to author guidelines
 
 * Download dataset from https://deepmind.com/research/open-source/open-source-datasets/kinetics/
 
-Please read this: https://arxiv.org/pdf/1808.01340.pdf
-
 
 REFS:
 https://www.ostechnix.com/20-ffmpeg-commands-beginners/
@@ -41,8 +39,10 @@ https://www.ostechnix.com/20-ffmpeg-commands-beginners/
 
 # import the libraries we need
 import os, sys
+from natsort import natsorted
 from tqdm import tqdm as tqdm
 import shutil
+from pytube import YouTube
 
 
 # Class begins
@@ -72,11 +72,22 @@ class KineticsDatasetManager(object):
 
 		# chek if destination exists. if it does, delete and create new one
 		if os.path.exists(self.destination_path):
+			user_input = input("Destination path '{}' already exists. Do you want to delete and recreate it? y or n: ".format(self.destination_path))
 			print("")
-			print("Destination path '{}' already exists. Deleting and re-creating...".format(self.destination_path))
-			print("")
-			shutil.rmtree(self.destination_path)
-			os.makedirs(self.destination_path)
+
+			if str(user_input).lower() == 'y':
+				print("Deleting and re-creating {}...".format(self.destination_path))
+				print("")
+				shutil.rmtree(self.destination_path)
+				os.makedirs(self.destination_path)
+
+			elif str(user_input).lower() == 'n':
+				print("Keeping '{}' as the default...".format(self.destination_path))
+				print("")
+
+			else:
+				print("Invalid {} option. Program exiting!".format(user_input))
+				sys.exit()
 
 		else:
 			os.makedirs(self.destination_path)
@@ -97,7 +108,82 @@ class KineticsDatasetManager(object):
 		-- To do --
 		[ref: https://github.com/ytdl-org/youtube-dl/issues/622#issuecomment-162337869]
 	"""
-	def download_video(self):
+	# provide range of downloads. Default is everything
+	def download_video(self,start_from=1,end_at=-1):
+
+		# lets make sure that the range we are providin
+		# this code needs cleaning. kindly make a pull request if you can help
+		def download_data_range(csv_location: str, dataset_type: str) -> dict:
+			with open(csv_location,"r") as opened_csv:
+				lines = opened_csv.readlines()
+				lines = [line.rstrip('\n') for line in lines]
+				lines = lines[1:] # trim out the label
+
+
+				# hold out dataset does not have labelling
+				# so we need to be careful about
+				if dataset_type!="holdout":
+					lines = [line.split(",")[0] for line in lines] # get all the labels
+
+					print("List of data set classes available for download")
+					print("-----------------------------------------------")
+					print("")
+					# now create a dictionary of the
+					# first element
+					indexing_holder = {}
+					start_point = end_point =  0
+					unique_element_counter = 1
+					for index_,element_ in enumerate(lines[:]):
+						
+						# variable to hold the last previous element
+						previous_element = lines[index_-1]
+
+						# however, if this is the first element, then previous = first
+						if index_ == 0:
+							previous_element = lines[0]
+
+						# if this curent element is the same as the previous one
+						# extend the end point
+						if lines[index_] == previous_element:
+							end_point = index_
+
+							# if we have reached the last element in the the array, 
+							# then set the stop point at this index
+							if index_ == (len(lines)-1): 
+								end_point = index_
+								print(unique_element_counter,". ",lines[index_])
+
+								# dictionary format: [id] ---> saves [class_name,index_start_point,index_end_point]
+								indexing_holder[unique_element_counter] = [lines[index_],start_point,end_point] # store the values in a dictionary
+								unique_element_counter +=1
+
+						# else if the curent element is different from the previous
+						# then let this index be the starting point of the new bath and
+						# let the prevoius index be the stop point of the previous batch 
+						else:
+							#print(lines[index_-1]," {} - {}".format(start_point,end_point ))
+							element_ = lines[index_-1]
+							end_point = index_-1
+							print(unique_element_counter,". ",element_)
+
+							# dictionary format: [id] ---> saves [class_name,index_start_point,index_end_point]
+							indexing_holder[unique_element_counter] = [element_,start_point,end_point]
+							unique_element_counter +=1
+							start_point = index_
+
+					# pretty printing
+					print("")
+
+					# return the dictionary
+					return indexing_holder
+
+				# -- to do: implement for when downloading holdout dataset
+				else:
+					sys.exit("holdout data set has not been implemented yet")
+							
+			# -- END --
+
+
 
 		if self.dataset_type == "train":
 			csv_location = "./dataset_splits/kinetics_600/kinetics_train.csv"
@@ -115,6 +201,71 @@ class KineticsDatasetManager(object):
 			sys.exit("Invalid dataset category type. Please enter [train,validate,test]")
 
 
+		download_data_range = download_data_range(csv_location,self.dataset_type)
+
+
+
+		"""
+		***********************************************************************************************
+		*************************** EVERY BIT OF CODE BELOW NEEDS SOME CLEANING UP ********************
+		************************ FOR NOW JUST MAKE SURE YOU ENTER THE RIGHT COMMANDS ******************
+		***********************************************************************************************
+		"""
+
+		# ask user to enter the range of the dataset they want to download
+		user_input = str(input("Please choose a range of the dataset classes you want to download, eg. 1 or 1-100: "))
+		print()
+
+		# now some validation
+		# trom out all spaces
+		user_input = user_input.replace(" ","")
+
+		# now check to see if a single number is given or a range is provided
+		user_input = user_input.split("-")
+
+		#print(user_input)
+
+		if len(user_input)>2:
+			sys.exit("Sorry. Please provide the correct range such as eg. 1 or 1-100. Program will exit now")
+
+
+		elif len(user_input)==1:
+			if isinstance(int(user_input[0]), (int)) is False or user_input[0]=='':
+				sys.exit("Sorry. Please provide the correct range such as eg. 1 or 1-100. Program exiting now")
+
+			user_input = int(user_input[0])
+
+			# ensure the right range
+			if user_input<=0 or user_input>len(download_data_range):
+				sys.exit("Sorry. Please provide the correct range such as eg. 1 or 1-100. Program exiting now")
+
+
+			start_from = download_data_range[user_input][1]
+			end_at = download_data_range[user_input][2]+1
+			print("Downloading videos of '{}' only.".format(download_data_range[user_input][0]))
+			#print(start_from, " --> ",end_at  )
+			print("")
+
+
+		elif len(user_input)==2:
+			if isinstance(int(user_input[0]), (int)) is False or isinstance(int(user_input[1]), (int)) is False or user_input[0]=='' or user_input[1]=='':
+				sys.exit("Sorry. Please provide the correct range such as eg. 1 or 1-100. Program exiting now")
+
+			from_class = int(user_input[0])
+			to_class = int(user_input[1])
+
+			# ensure the right range
+			if from_class>=to_class or from_class<=0 or to_class<=0 or from_class>len(download_data_range) or to_class>len(download_data_range):
+				sys.exit("Sorry. Please provide the correct range such as eg. 1 or 1-100. Program exiting now")
+
+
+			start_from = download_data_range[from_class][1]
+			end_at = download_data_range[to_class][2]+1
+			print("Downloading videos from class '{}' to '{}'".format(download_data_range[from_class][0],download_data_range[to_class][0]))
+			#print(start_from, " --> ",end_at  )
+			print("")
+
+
 		#number may not be right but just give us a rough estimate	
 		video_counter = 0
 
@@ -128,8 +279,8 @@ class KineticsDatasetManager(object):
 			lines = lines[1:]
 			sp = "./"
 
-			# loop through the lines to download the video
-			for line in tqdm(lines):
+			# loop through the lines to downlad the videos
+			for line in tqdm(lines[start_from:end_at]):
 				coumn = str(line).split(",")
 
 				# there is no label for holdout data
@@ -144,6 +295,7 @@ class KineticsDatasetManager(object):
 					start_time = coumn[2]
 					
 
+				
 				# create the directory according to the label name
 				dir_name = str(data_lable).replace(" ","_") # replace space with underscore
 				dir_name = os.path.join(self.destination_path,dir_name) # make it an absolute path
@@ -159,9 +311,14 @@ class KineticsDatasetManager(object):
 
 				# create the youtube link
 				youtube_link = "https://www.youtube.com/watch?v="+youtube_id
-
+				
+				
 				# use youtube-dl and ffmpeg to download videos
 				os.system("ffmpeg -hide_banner -ss "+start_time+" -i $(youtube-dl -f 18 --get-url "+youtube_link+") -t 10 -c:v copy -c:a copy "+vid_path)
 				video_counter +=1
 				print(video_counter, " videos download")
+				
+				
+				#print("vide path: ",vid_path)
+
